fix(lint): split ui/routines/filter[_tests].rs under the 500-line gate
